non-functional real-time search implementation
Doesn't work, not sure why. Seems nobody has ever tried to create a custom real-time search command before. At least the haven't put anything on the interwebs about it.

The command does get called by splunk, but just spits out:

    Error in 'influx' command: This command is not supported in a real-time search

The call metadata looks like:

    {"action":"getinfo","preview":false,"streaming_command_will_restart":true,"searchinfo":{"args":["db=bitcoin","select","*","from","trades"],"raw_args":["db=bitcoin","select","*","from","trades"],"dispatch_dir":"/opt/splunk/var/run/splunk/dispatch/rt_1497159365.72","sid":"rt_1497159365.72","app":"search","owner":"admin","username":"admin","session_key":"WN67iFWzLHbIcapw2T2^qxcPQuDeAcAir0XSKTozu4VeVltasnovqmss7H_wrAHPVl0Q2VxP7zakJSP2yYCFXEFQcveFdF6271AwEqLNx98Quiol9FRPBbCkCta","splunkd_uri":"https://127.0.0.1:8089","splunk_version":"6.6.1","search":"%7C%20influx%20db%3Dbitcoin%20select%20%2A%20from%20trades","command":"influx","maxresultrows":50000,"earliest_time":"0","latest_time":"0"}}

My guess is that we're not responding to the 'getinfo' with the right metadata. But I have no clue what that "right metadata" is.